requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
antlr4-python3-runtime==4.8
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy==0.7.0
cachetools==5.0.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.2
cryptography==37.0.1
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
einops==0.4.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.1
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.3.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.44.0
h5py==3.7.0
hydra-core==1.1.2
idna==3.3
imageio==2.18.0
importlib-metadata==4.11.3
importlib-resources==5.2.3
ipykernel==6.9.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.10.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.2
lxml==4.9.1
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8
notebook==6.4.12
numpy==1.23.1
nystrom-attention==0.0.11
oauthlib==3.2.0
omegaconf==2.1.2
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.1.2
pkgutil_resolve_name==1.3.10
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.6.1
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.2.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.5.10
setproctitle==1.2.3
setuptools==61.2.0
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.4.22
timm==0.6.5
tinycss2==1.1.1
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.8.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.1.1
urllib3==1.26.11
wandb==0.12.15
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
wheel==0.37.1
widgetsnbextension==3.6.1
yarl==1.7.2
zipp==3.8.0
